timeline: move the back-pagination code into the timeline/pagination.rs file
No changes in behavior, just pure code motion.
